rag-backend/requirements.txt
sentence-transformers
torch
numpy
qdrant-client
fastapi
uvicorn
python-dotenv
openai

